requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-community
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
beautifulsoup4
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
sentence-transformers
SQLAlchemy
mysql-connector-python
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
gradio
openai
langchain_nvidia_ai_endpoints
crewai
crewai_tools
langchain_experimental
langchain_neo4j
lamini
langgraph
langsmith
cassio
tiktoken
langchainhub